+++ packages/server/src/server.spec.ts
@@ -80,7 +80,7 @@ describe("setupApp", () => {
 
         expect(actual.header["content-type"]).toBe("application/json; charset=utf-8");
         expect(actual.statusCode).toBe(200);
-        expect(unpackPayload(actual.body)).toBe("Expected result");
+        expect(unpackPayload(actual.body)).toEqual(expect.objectContaining({ data: "Expected result" }));
     });
 
     it("responds with index page w/ staticDir and GET to /", async () => {