+++ packages/shared/src/transport/index.ts
@@ -8,6 +8,7 @@
 export type { NamespaceMapping } from "./NamespaceMapping";
 export type { RemoteFunction } from "./RemoteFunction";
 export type { RequestPayload } from "./RequestPayload";
+export type { ResponseError } from "./ResponseError";
 export type { ResponsePayload } from "./ResponsePayload";
 export { Serialization } from "./Serialization";
 export type { TransportFunction } from "./TransportFunction";