+++ packages/test/test/__data__/faas-modules/echo.ts
@@ -1,0 +1,5 @@
+
+// @service()
+export const foobar  = async (echo: unknown) => {
+    return echo;
+}