+++ Release-Notes.md
@@ -12,6 +12,10 @@ Release notes follow the [keep a changelog](https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Added
+
+- @service functions throwing on the server ensure the client promise is rejected.
+
 ### Changed
 
 - magellan-server has tslib as production dependency to ease integration in custom servers.